Moves Varnish classes into Addon/Varnish. Fixes tests.
